blink/web_tests/fast/forms/week-multiple-fields/week-multiple-fields-ax-aria-attributes.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../resources/multiple-fields-ax-aria-attributes.js"></script>
</head>
<body>
<input id="test" type="week" value="2012-W10" aria-label="TestLabel" title="TestTitle">
<script>
var testInput = document.getElementById('test');

test(function(t) {
    testInput.focus();
    assert_equals(focusedElementDescription(),
                  'Week TestLabel, --, "10", TestTitle, intValue:10, range:1-53');
    eventSender.keyDown('\t');
    assert_equals(focusedElementDescription(),
                  'Year TestLabel, ----, "2012", TestTitle, intValue:2012, range:1-275760');
}, 'Non-empty value');

test(function(t) {
    eventSender.keyDown('Backspace');
    assert_equals(focusedElementDescription(),
                  'Year TestLabel, ----, "", TestTitle, intValue:0, range:1-275760');
    eventSender.keyDown('\t', ['shiftKey']);
    eventSender.keyDown('Backspace');
    assert_equals(focusedElementDescription(),
                  'Week TestLabel, --, "", TestTitle, intValue:0, range:1-53');
}, 'Empty value');

</script>
</body>
</html>
